sections.component replaced with correct file
